requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
python-multipart
psycopg2-binary
httpx
pydantic[email]
